Added Chatbox.jsx and timer to ChatHistory.jsx
